requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
Django==5.1.3
django-appconf==1.0.6
django-compressor==4.5.1
django-crispy-forms==2.3
django-libsass==0.9
libsass==0.23.0
python-dotenv==1.0.1
pytz==2024.2
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.2
tzdata==2024.2
